fixes scrapper run, findAll (byTime) method in JdbcLinkDao and refactor structure of project
