+++ scrapper/src/main/java/edu/java/scrapper/controller/chatApi/ChatController.java
@@ -1,6 +1,6 @@
 package edu.java.scrapper.controller.chatApi;
 
-import edu.java.scrapper.service.tgChatService.TgChatService;
+import edu.java.scrapper.service.daoService.TgChatService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
