+++ scrapper/src/main/java/edu/java/scrapper/controller/linksApi/LinksController.java
@@ -4,8 +4,8 @@ import edu.java.scrapper.controller.request.AddLinkRequest;
 import edu.java.scrapper.controller.request.RemoveLinkRequest;
 import edu.java.scrapper.controller.response.LinkResponse;
 import edu.java.scrapper.controller.response.ListLinkResponse;
-import edu.java.scrapper.domain.link.Link;
-import edu.java.scrapper.service.linkService.LinkService;
+import edu.java.scrapper.domain.dto.Link;
+import edu.java.scrapper.service.daoService.LinkService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
