+++ scrapper/src/main/java/edu/java/scrapper/domain/dto/Chat.java
@@ -1,4 +1,4 @@
-package edu.java.scrapper.domain.tgChat;
+package edu.java.scrapper.domain.dto;
 
 import java.time.OffsetDateTime;
