+++ scrapper/src/main/resources/application.yml
@@ -16,7 +16,7 @@ spring:
   application:
     name: scrapper
   datasource:
-    url: jdbc:postgresql://postgresql:5432/scrapper
+    url: jdbc:postgresql://localhost:5432/scrapper
     username: postgres
     password: postgres
   liquibase:
