+++ scrapper/src/test/java/edu/java/scrapper/domain/dao/jdbc/JdbcTgChatTest.java
@@ -1,7 +1,8 @@
-package edu.java.scrapper.domain.tgChat;
+package edu.java.scrapper.domain.dao.jdbc;
 
 import edu.java.scrapper.IntegrationTest;
-import edu.java.scrapper.domain.tgChat.jdbcImpl.JdbcTgChatDao;
+import edu.java.scrapper.domain.dto.Chat;
+import edu.java.scrapper.domain.dao.jdbc.JdbcTgChatDao;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
